Implement basic geocoding.
Removed the old build.rs implementations.
Now, at runtime, we download the required files from geonames.org.
A command line parameter specifies which countries we will need
to download. Downloads are only done once, which makes the whole
process fast.

The actual reverse-geocoding lookup, however, is very slow. So instead
of applying it to all parts of the point, we probably only want to do
it to enough points to get a stage description.