Remove maud, use rinja instead.
Reason: Maud allocates a string for every partial, and we are going to have a
lot of them. While it would probably be ok, why risk it? Rinja renders to the
output stream in one allocation. Rinja is a fork of Askama and is used as the
template library by docs.rs, so we are on safe ground with it.